Add role and permission

+++ app/Http/Controllers/Admin/UserController.php
@@ -5,29 +5,40 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\User;
 use Illuminate\Http\Request;
+use Spatie\Permission\Models\Role;
 
 class UserController extends Controller
 {
+    function __construct()
+    {
+        $this->middleware('can:create user', ['only' => ['create', 'store']]);
+        $this->middleware('can:edit user', ['only' => ['edit', 'update']]);
+        $this->middleware('can:delete user', ['only' => ['destroy']]);
+    }
+    
     public function index()
     {
         $search = request('search', null);
         $users = User::when($search, function($user) use($search) {
             return $user->where("name", 'like', '%' . $search . '%')
             ->orWhere('id', $search);
-        })->paginate(10);
+        })->paginate();
+        $users->load('roles');
         return view('admin.user.index', compact('users'));
     }
 
     public function create()
     {
-        return view('admin.user.create');
+        $roles = Role::pluck('name', 'id');
+        return view('admin.user.create', compact('roles'));
     }
 
     public function store(Request $request)
     {
         $input = $request->only('name', 'email', 'password');
         $input['password'] = bcrypt($request->password);
-        User::create($input);
+        $user = User::create($input);
+        $user->assignRole($request->role);
         return redirect()->route('admin.user.index')->with('success', 'A user was created.');
     }
 
@@ -38,7 +49,9 @@ class UserController extends Controller
 
     public function edit(User $user)
     {
-        return view('admin.user.edit', compact('user'));
+        $roles = Role::pluck('name', 'id');
+        $userRole = $user->getRoleNames()->first();
+        return view('admin.user.edit', compact('user', 'roles', 'userRole'));
     }
 
     public function update(Request $request, User $user)
@@ -48,11 +61,15 @@ class UserController extends Controller
             $input['password'] = bcrypt($request->password);
         }
         $user->update($input);
+        $user->syncRoles($request->role);
         return redirect()->route('admin.user.index')->with('success', 'A user was updated.');
     }
 
     public function destroy(User $user)
     {
+        if(auth()->id() === $user->id) {
+            return back()->withErrors('You cannot delete current logged in user.');
+        }
         $user->delete();
         return redirect()->route('admin.user.index')->with('success', 'A user was deleted.');
     }
